Add files via upload
Make a program to read a student's name and the three grades he got in the three quarters of the year 
(the first quarter is worth 30 and the second and third are worth 35 each). At the end, display the student's final grade for the year. It also tells you whether the student passed (PASS) or failed (FAILED) and, if not, how many points are missing 
for the student to obtain the minimum to pass (which is 60% of the grade). You should create a Student class to 
solve this problem.